requirements.txt
python-telegram-bot[job-queue]>=21.0
openai>=1.0
python-dotenv>=1.0
feedparser>=6.0
aiohttp>=3.9
beautifulsoup4>=4.12
faster-whisper>=1.0
numpy>=1.26
twikit>=2.0
requests>=2.31
youtube-transcript-api>=1.0
duckduckgo-search>=8.0
anthropic>=0.40
